Created and implemented TDList and TDItem along with the add feature within the main application.
Implemented a TDResult object along with a public Enum for result returns and error handling.

The add item allows you to either a) use the list id or b) use the list name

Application can export a serialized file to store data when closing and will load if the file is present.